WIP Fix deletion inside a structural node at start/end of document
Logic assumed that if you had a collapsed selection near the start
or end of the document that you didn't want to delete anything,
however this is only true for content nodes on their own (e.g.
the first/last paragraph). Paragraphs in structural nodes (e.g.
lists) should be unwrapped).

Bug: 50250
Change-Id: I1ca31b2c8dbc4a01d8f7363895e7593eef10d702