Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

I still think the behaviour is not very clear. Your comment at the top says I can instantiate with or without "size details", but there are two components to that: size and aspect ratio. Can I define size but not AR, AR but not size, both? Is AR assumed to be from the current size if not defined?

setDimensions lets me set any size if AR is null, but then calls setOriginalDimensions which will lock the AR? Is a fixed AR required or configurable?

Here's how I see it should be laid out:

 onWidthChange {
   validate(this.widthInput);
   if(isValid && ARFixed && !this.changing) {
     this.changing = true;
     this.heightInput.setHeight(w*AR);
     this.changing = false;
   }
 }

(NB height is set before validating which happens in onHeightChange, and then it becomes clear to the user why height validation failed)

That completely handles the user input cases, there is no need for width/height properties or setDimensions.

In the case of programmatic setting, either you aren't changing the AR, in which case you can just call widget.setWidth() or widget.setHeight(), or you are, in which case you call setOriginalAspectRatio first. If the AR was never fixed, you can call setWidth then setHeight, or if you really want, have a setDimensions(w,h).

Patch-set: 8
Label: Code-Review=-1