Update patch set 8
Patch Set 8:

I think I may not have been clear in the documentation of how I see the widget working (and I will fix that) but I want to make sure that is the problem rather than the actual way the widget works.

The inner values of this.width and this.height preserve 'raw values (doubles) of the calculations, and the inputs themselves are then rounded. This has two main benefits the way I see it:

1. It saves the calls to input.getValue() each time a calculation is done.
2. It allows for better mathematical validation on whether the values changed or not, because it saves them in the original non-rounded state.

Most OO.ui widgets use internal values that sync to the input, including InputWidget, so I didn't think it was a completely wrong way to go -- especially since in this case, where the mathematical values benefit from having the actual raw value. So, calculating this.height = (this.width)1.345*(this.aspectRatio)1.3333 and then rounding the result will give a more accurate rounded answer than if we multiple with already rounded width/height values.

As for the behavior itself, the intention is to allow users to instantiate the widget with given width/height + originalWidth/originalHeight (aspect ratio is then calculated from it) or to supply those after instantiation, like in the case of the Media Edit dialog.

Let's go at it from the different direction; here's how I saw the widget being called and used in the Media Edit Dialog:

(in .initialize)
this.sizeWidget = new ve.ui.MediaSizeWidget( { '$': this.$ } );

(in .setup)

this.sizeWidget.setOriginalDimensions( width/height from API );
this.sizeWidget.setDimensions( from current image dimensions )

So, everything else is fallback options. 

1. If the originalDimensions aren't available, I will use the current dimensions to calculate the aspect ratio. *But* I only want to use the first current dimensions for two reasons:
    a. I can get a consistent aspect ratio from them.
    b. I can still have 'revert to original dimensions' button that simply returns the image to the original size you loaded it with - even if the user "played around" with the numbers and "lost" the original value while editing (before saving).

2. I wanted to make sure the widget doesn't fail on 'fringe' events. For example, if my page (in wiki or outside, since this is a core widget) has an image that only has one dimension value: <img src="whatever.png" width="100" /> if I find originalDimensions, the widget will use that to calculate aspect ratio -- but if it doesn't have originalDimensions, then as a fallback I will assume 1:1 ratio. This fallback, imho, is perhaps not as logical, but after consulting with James about it, it seemed like a good-for-now solution.

That means that I still have to use originalDimensions and this.width/this.height.

I gave the option to instantiate the widget with prefilled conditions because I thought it might be needed for some extension. In that case, again all we need to supply are 
1. Current width/height + originalDimensions
2. Current width/height (and original dimensions are taken from the first of these values)
(I do notice that #2 has a bug in the code, and will fix it)

This seemed to me to be good behavior; we always either go by current width/height or current+original. All other things are calculated accordingly.

So what you pointed out as weird behavior are actually fallback cases. 

I understand I need to document them better, but is it really a case of wrong behavior -- or does this explanation make it clearer?

Patch-set: 8
Reviewer: Mooeypoo <798@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0