Update patch set 2
Patch Set 2:

Timo's comment from this change in the other repo before it was moved:

"I'm for separating concerns, but I don't think calling an .onSomething method directly is a good direction to go in.

It's may perhaps seem slightly overkill in this specific case, but I think it's worth it in this case to abstract it just a bit further to be more explicit (e.g. don't call onSomething, and don't pass the event object, but turn it into a method that you call from both places).

Earlier today I watched a talk that reminded me of it and does a good job of explaining it (http://www.youtube.com/watch?v=nZihjH6_Qns#t=30m10s)."

Patch-set: 2