Collapse block slugs and expand on hover/focus
By default collapse the height of block slugs to about half the height
and expand on hover or focus with CSS. Show a dotted outline on hover
to make it clear you are inserting a new paragraph.

As the :focus psuedo class doesn't work on contenteditable's children,
toggle a class in the slug detecting code of SurfaceObserver.

Bug: 47790
Change-Id: Iab25fc6d5e21d0014482913e148c53b32d9652c4