Update patch set 4
Patch Set 4: Code-Review-1

Playing locally with this in the standalone demo (page=default/aliens) in Chrome shows it as very jerky. Something seems to be stuck in a loop whenever I get near it. No transitions are visible. The block slog is expanding/collapsing rapidly and disappearing partially as well.

Patch-set: 4
Label: Code-Review=-1