Update patch set 2
Patch Set 2:

The documentation helps! Were the *properties.js files regenerated with unicodejs-properties.py ? If so, I guess (1) we should note so in the commit message, and (2) we should probably make the python script prepend an "Automatically generated by ... " comment to them (at line 45 of the script).

Patch-set: 2
Label: Verified=0