Allow dynamic limit to maxDimensions
Make sure that the ability to limit to maximum dimensions is dynamic
and can change. This is beneficial for when the image type changes
to thumbnail (which requires a limit) or to anything else (that can
allow for any size without limit.)

Also, allow for a forced revalidation of the current size widget
dimensions in case limits were applied or removed.

Change-Id: I362c7c378c5ccd7a83495b03e4429c9b93c1e9ee