Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

The margin set on the focussed block slug seems a bit arbitrary. Between focussing the block slug and typing (at which point it is replaced with a regular paragraph containing the same text) it will flash due to margin/padding change depending on the margin specified for paragraphs in the stylesheet. In the standalone demo we don't specify any paragraph margins and inherit the browser defaults. In MediaWiki that will be different and elsewhere different as well. It doesn't flickr for me by default because the margin we use is close enough to the default for HTML5 pages in WebKit/Blink. We should probably either apply the same margin to all paragraphs or do it the other way around (let it inherit and only override when we shrink it, not to let it grow). Let's go with the latter since we don't want to change the looks of the page, that's integration responsibility and will improve look-alike ness between editor and view.

Patch-set: 7
Label: Code-Review=-1