ve.ce.Node: Don't use padding or height in the slug transition
These properties can unexpectedly transition in the wrong
direction due to us not being (and not wanting to be) in control
of these for regular paragraphs.

Increasing font-size slightly to make up the absolute difference
in size of the hover state with the height and padding gone.

Increasing transition duration in attempt to smoothen it a bit.

Change-Id: I3ee215f869873420df7017b090c92ee37501746a