Update patch set 9
Patch Set 9:

(3 comments)

Code looks fine, untested. I'm a bit suspicious of how the emission of events happen. Event handlers (the DOM ones, not EventEmitter ones) aren't exactly free. It seems a bit odd to have all these different mixins bind a dbclick event on their object (focusable shield, phantom protect shield, draggable target etc.) when most of these will remain unused, and in the most common case a CE node will have all of the above and only the top most will actually work. Maybe the evnet handling could be handled by the ClickableNode entirely and have some kind of method, config or property that will declare what to bind it on?

e.g. something like

 ve.ce.FooNode
   mixin FocussableNode
   mixin ClickableNode

   node.bindClickTarget( node.$highlights );

Or more plain:

 ve.ce.FooNode
   mixin FocussableNode
   mixin ClickableNode

   node.$highlights.on( 'dblclick', function () { node.emit( 'dblclick' ); } );

That way it is the responsibility of the CE node to tell it which one(s) to listen on, and the other ones won't have semi-arbitrary event handlers for "dblclick" which is manageable now, but I fear isn't a model that will scale very well for other things, and would be impossible to compete with for plugins if our pattern is to emit them from inside other mixins.

Anyway, just rough ideas. Maybe it's better the way you have it, happy to discuss it if you like.

Patch-set: 9