Update patch set 9
Patch Set 9:

I like the idea of not binding to things that will just be buried.   Perhaps per your suggestion I could determine all of the possible mixin cases and then suggest which element to bind to ?   For example if you mixin relocatable node, the image element used for dragging is the element focused when clicked, so the dbl click would need to be bound to it opposed to other elements in the overlay.   The main issue really is future proofing

Patch-set: 9
Label: Verified=0