Fix JS error when pressing Enter with selection across multiple paragrahps
When the selection spanned multiple ContentBranchNodes, handleEnter would
first get the node where the selection started, then perform a removal,
then do stuff with the node. However, the removal detaches the node in
this case (because it gets merged with another node, so that part of the
tree is rebuilt), which results in errors later on when we assume that
the node has a non-null parent.

Instead, perform the removal first, then grab the node.

Bug: 62235
Change-Id: Id6266eaa24c48bcbca3279fb8b6ca0e1ba18cce0