Fix ContentBranchNode echo suppression at the ends of text nodes
Rename selectionInsideOneBranchNode() to selectionInsideOneLeafNode(),
and use selectNodes() to check if the selection only covers one node.

This is needed because we don't just care about the selection being
within a CBN, we also want to steer clear of selections from a
text node across an inline node into another text node, those are
not "simple".

Change-Id: I85fe939d3709e65e7094cbc55f29bdeb673df15a