[WIP^2] Refactor Scalable and MediaSizeWidget
Scalable is now a dm calculation engine, responsible for all the
size and constraint calculations of a scalable object. MediaSizeWidget
takes a scalable object to interact with the user and process the values
supplied.

**
This commit is up on gerrit for backup purposes since my VM is
unpredictable. The commit requests you not look at it. It's hedious.
Seriously. Move along. Nothing to see here just yet.
***

Change-Id: Ie7ca3bbbe5d0287c7ce17d296c1827b8213ea950