Refactor Scalable functionality
This fix transformed scalable into a calculation engine that is
delivered by resizable objects. The scalable object is responsible
for all calculations and adjustments that relate to size for the
given node.

As a result, everything that relates to size calculation has been
touched.

On top of that, media size widget has been adjusted to work with
scalable object and also display the dimensions through the
dimensions widget.

This is a breaking change, fixed in MW by change
I017a017924f544cc8bc9b7d8245335759ae0e890

Change-Id: Ie7ca3bbbe5d0287c7ce17d296c1827b8213ea950