[WIP] Followup 1b9c077: debounce toolbar's onSurfaceViewKeyup
This event handler measures the toolbar and the position of the
cursor, which turns out to be fairly expensive on Chrome. Measuring
the cursor position on every keypress slows down typing a lot,
and it's not really necessary to reevaluate this one milisecond
after we concluded that the cursor wasn't behind the toolbar.

In Firefox, however, this isn't slow at all.

TODO: Instead of simply debouncing, we should always remeasure
in certain cases like if the previous keypress resulted in a
scroll adjustment or if the keypress was an arrow key.

Change-Id: I5b3ffd40115262e67176cc9f6978ea74c5a69a1a