Protect against invalid/absent <base> in ve.resolveUri()
If the document has no <base> and has nowhere to inherit a base URI from
(this is the case for documents created with real document creation functions
rather than the iframe trick), things like node.href will return empty string (!!)
if the href is relative. This is crazy, so detect this case and return the
original value instead.

Change-Id: If1dd4421d1eb0e73e3b86821f4e0aa5a82c851a2