Pass SurfaceFragments, not Surfaces, to Dialogs
Dialogs should only be interested in the model. The also need to know
the range that existed when the dialog was opened as it will be cleared
when the parent surface is blurred.

Add getSelectedNode to SurfaceFragment as dialogs may need to know
which node was 'focussed' when the dialog was opened.

Change-Id: Idd78bf9e4d50635b707bacce06f6c14babcf33a8