Adjust range manually in isolateAndUnwrap()
Relying on SurfaceFragment's magical updates will cause the range to expand
to cover the inserted separators if those separators are directly adjacent
to the range, but we need the range to keep covering exactly what's being
isolated and no more, so we need it to not grow.

This range growth doesn't usually happen, because the typical situation is:
[</list><list>]<li>|<p>Foo</p>|</li>[</list><list>]
(i.e. the list is being split but the listItem is not), but it's possible to get:
[</list><list>]<li>|<p>Foo</p>|[</li></list><list><li>]<list><li>...
when converting a list item that contains a nested list, in which case the
listItem does get split.

Because there's no .setRange() method in SurfaceFragment (yet?), this
uses some hacks to set the range.

Bug: 63209
Change-Id: Iec920dddecf501eb6fb8b2375b5af37bf9555cc5