Exclude insertions in ve.dm.SurfaceFragment#isolateAndUnwrap
In cases where the separators we insert are adjacent to the range
we're isolating, we don't want the range to grow to include these
separators, as that breaks the subsequent unwrap.

This range growth doesn't usually happen, because the typical situation is:
[</list><list>]<li>|<p>Foo</p>|</li>[</list><list>]
(i.e. the list is being split but the listItem is not), but it's possible to get:
[</list><list>]<li>|<p>Foo</p>|[</li></list><list><li>]<list><li>...
when converting a list item that contains a nested list, in which case the
listItem does get split.

Bug: 63209
Change-Id: Iec3fecccf5d4ad545164f36ee73b7285f29c0e55