Small validation fixes to scalable
There was an issue where scalable dimensions may have been set as
{ width: undefined } which made the tests in most 'set' methods
pass probperly, because this wasn't an empty object. However,
practically speaking this is an empty dimensions object, and it
should be considered as if it isn't valid.

This wrecked havoc on the calculations done with the aspect ratio,
and affected 'default dimensions' most of all.

Changes made:
* Added a 'isDimensionsObjectValid' method to ve.dm.Scalable that
  tests dimension objects specifically to see if they are valid
  before update is done to any of the stored dimension values.
* Added originalSizeChange event to ve.dm.Scalable and a listener
  in ve.ui.MediaSizeWidget to make sure that the 'full size'
  button is only enabled when original dimensions are acquired,
  even if the API response is slow or was completed after the
  widget is already loaded.
* Fixed small typos in setDefaultDimensions in ve.dm.Scalable

Change-Id: I1242e8dadca4f2cf2737c366034a77edc5ce308e