Use more standard format for displaying keyboard shortcuts
This affects both the tooltips on tools and the help dialog.
"CTRL + SHIFT + N" (uppercase keys, with a plus, additional spaces)
is just plain ugly.

Let's instead use "Ctrl+Shift+N" (titlecase keys, with a plus, no
spaces). This is the format used consistently throughout the UI of
Windows 7 (and probably other versions too, but I don't have them here
to check) and Ubuntu 12 (default theme), as well as native
applications on these systems. We already override the default
behavior for Macs, that is not changed.

MediaWiki core uses "ctrl-shift-n" format (lowercase keys, with a
dash, no spaces) for its accesskeys tooltips, that could be considered
as well for consistency, but I think it looks a bit silly.

Change-Id: Ifd1e8e01e5bee5bf7bc82731d5bea27794c6dfb9