Use more standard format for displaying keyboard shortcuts
Before: http://i.imgur.com/kSEagTj.png
After: http://i.imgur.com/XaJhrUX.png

This affects both the tooltips on tools and the help dialog.
"CTRL + SHIFT + N" (uppercase keys, with a plus, additional
spaces; monospaced font in the dialog) is just plain ugly.

Let's instead use "Ctrl+Shift+N" in normal font (titlecase keys, with
a plus, no spaces). This is the format used consistently thorough the
UI of Windows 7 (and probably other versions too, but I don't have
them here to check) and Ubuntu 12 (default theme), as well as native
applications on these systems. (Windows 7 does sometimes use the
spaced version as well, but never in the contexts we have here.) We
already override the default behavior for Macs, that is not changed.

The HTML5 spec for <kbd> also uses this format in its example:
<http://www.w3.org/TR/html5/text-level-semantics.html#the-kbd-element>.

MediaWiki core uses "ctrl-shift-n" format (lowercase keys, with a
dash, no spaces) for its accesskeys tooltips, that could be considered
as well for consistency, but I think it looks a bit silly.

Change-Id: Ifd1e8e01e5bee5bf7bc82731d5bea27794c6dfb9