Stage slugs
As soon the cursor is placed in a slug, stage the creation of
a real paragraph.

If the the cursor is moved out of the slug and it is still empty
then pop the staging stack.

If the staged paragraph is added to, commit the staging stack.

As part of this slugs are now CE=false to reflect the fact they
don't exist in the model until the are clicked (not just selected
over). This avoids strange behaviour with selections that only
start xor end in a slug.

Future to-do:
* Fix or kill animations as this breaks them
* Possibly:
  + Fix existing bug with left/right arrowing into a slug
  + Fix up/down arrowing into a slug now they're CE=false
* Kill getSlugAtOffset, and any other code which detects for the
  presence of a block slug.

Bug: 65052

Change-Id: Ic857215654eaf5ee32e19d7db1d83185f561f783