Update patch set 1
Patch Set 1: Code-Review+1

Seems like a good habit to not clash these either way. In theory this would allow one to instantiate desktop/mobile mediawiki/non-mediawiki targets on one HTML page (and even multiple of each). If the classes are generically overwritten like this, that is not possible.

We do the same in mw.ViewPageTarget.css (overwrite for the target specifically, not redefine the class in general).

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1