Update patch set 3
Patch Set 3:

The huge amounts of orange in your screenshot are caused by a browser bug where it doesn't like blocks inside of inlines. Currently we are treating all extensions as inline which I have a separate fix for.

In general, yes we won't the 'perfect' box selection we have once the shiled is rendered, but we will get a selection.

Patch-set: 3
Label: Verified=0