Update patch set 2
Patch Set 2:

If we're unable to test this I recommend going with just binding to 'paste' (which is what we always did in non-msie browsers and will not be able to break anything as it will not change anything). That reduces the risk to MSIE which defer it to a todo item on the list of IE support things to figure out.

Patch-set: 2
Label: Verified=0