Update patch set 2
Patch Set 2:

No, it's not. The same applies to ve.msg, that's inside the method, not in the global scope. But anything exposed by a module should be immediately useable. When loading the module this file belongs to (and its declared dependencies), can this class be instantiated? If not, is that something that is snafu (because it's the same missing link), or an additional missing link.

Patch-set: 2