Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Migrating debugBar to OO.ui.Element is a good change, but the other changes are all unnecessary or wrong.

Most of your changes involve passing an already-constructed DOM element to jQuery. In that case it doesn't matter if you use $() or this.$(), because jQuery gets the correct document from the DOM element already. There is only a difference between this.$() and $() when it comes to construction of DOM nodes (subtle difference) and resolving CSS selectors (obvious difference); in both cases, the argument is a string.

There were a few places where you changed $('selector') to this.$('selector'). This is obviously a functional difference. In most places where $() is used, especially places that deal with the global overlay, it's a conscious choice to use the global html/body element, because the global overlay is global. We may want to revisit that (right now none of our Targets are properly embeddable, they all use the global document), but right now that's how things are done.

Patch-set: 2
Label: Code-Review=-1