Update patch set 1
Patch Set 1:

How does this fix the bug? Is there a moment where text is selected rather than a focusable node, like between the removal of the old node and the insertion of the new one? Also, wouldn't this break programmatically moving the selection from a focusable node to text? (Do we even do that anywhere?)

Patch-set: 1