Update patch set 1
Patch Set 1:

Wait, you're saying that if you just cursor off a focusable node, there is code *other than this code* that causes the surface to be focused? That really surprises me.

Is the fundamental problem that .focus() isn't idempotent and has negative effects if called on an already-focused surface?

Patch-set: 1