Update patch set 6
Patch Set 6: Code-Review-1

This breaks hiero nodes. They consist of a <table> and a <span> (the span only contains a newline, and only if Tidy is disabled or something -- it doesn't appear in production), both of which measure 0x0 but the outer div does have nonzero dimensions. Replacing the shield in the div (which had dimensions, and wasn't 100% width due to display:table;) with shields in the table and the span (which are both 0x0) means hiero nodes have no highlights.

Patch-set: 6