Update patch set 1
Patch Set 1: Code-Review-1

The same issue (selecting </p>) needs to be fixed in a lot of other cases as well, such as when you use the image dialog to change an inline image to a block image, or when you insert a reference list in the middle of a paragraph. These are basically all the same issue and will have to be solved in basically the same way, so I think you should fix it all in one place in the same way, and ideally do it in a way that's a bit less dirty.

Relevant bugs: https://bugzilla.wikimedia.org/show_bug.cgi?id=65706 https://bugzilla.wikimedia.org/show_bug.cgi?id=65716

Patch-set: 1
Label: Code-Review=-1