[BREAKING CHANGE] Show comments inline to the user
Add a CommentNode and allow for <!--comments--> to be rendered
as nodes in VisualEditor. Comments are now inspectable nodes
with the ability to view them, edit and add new ones.

For now, only show comments inside of content branch nodes;
comments in structural positions can often be shown, but sometimes
they can't be (if they're in the middle of a table for instance)
and detecting that difference isn't easy.

Breaking change: Introduces a new comment node which may break tests.

Bug: 49603
Change-Id: I05fcccdeb881d1af245077d2991bff4458031340