Update patch set 17
Patch Set 17: Code-Review-1

Hm.. I'm not sure if this commit was different first, or that there is a different intend than I think, or maybe there is the same indent but implemented in a confusing way. Why are we creating separate 'dist' modules and demos? Why not simply have the buildloaders always use the dist? We do that in oojs-ui as well. No need for separate module definitions, demo target or 'baselibs'.

Patch-set: 17
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1