Update patch set 9
Patch Set 9:

So it looks like you're moving it the context to right above the toolbar. This works in the MW integration because the toolbar is at the top of the screen there, but in any integration where that is not the case (like in the mobile demo in VE core), this will break.

Patch-set: 9