Update patch set 5
Patch Set 5:

If we're going to support font size I guess we'd need something like this, however I would support it differently, perhaps as a "textSize" annotation with a size/level property: -1, 0, 1, which then renders as <small>, '', <big>. Maybe also have the annotation self-destruct when size==0.

That way bigger/smaller are just bounded size++/size-- operations, and we can extend to more sizes in the future.

Patch-set: 5