Update patch set 5
Patch Set 5:

In the same way we 'support' nested sup/sub but if you apply fresh annotations we clean up to just one or the other, we should do the same with font size tags.

Also you say <big><big> is used frequently, but it seems like a pretty rare edge case to me.

In fact how often are big/small used at all in articles. Is it something we want to encourage?

Patch-set: 5
Label: Verified=0