Update patch set 20
Patch Set 20: Code-Review-1

The scroll height issue appears to be a rounding error. getBodyHeight on a fragment inspector returns 150.200px for this.form.$element.outerHeight( true ), maybe because of the em padding? Using Math.ceil appears to fix the problem, but it may need to be used in every place we calculate a height to prevent future issues.

Patch-set: 20
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1