Update patch set 2
Patch Set 2:

Indeed. At the moment all of VisualEditor uses the apex theme of OOjs UI, so it'd make sense to not have the extension of that in VisualEditor suddenly use Agora. This is testified by the fact that the "alert" icon in effect in VisualEditor from OOjs UI isn't orange either (because we load the apex theme).

Patch-set: 2
Label: Verified=0