Update patch set 1
Patch Set 1: Code-Review-1

If this change is being made in the name of consistency, it fails, because we don't use this grey elsewhere.

This change also fails the product objective of making the comment extremely high-profile and noticeable.

Patch-set: 1
Reviewer: Jforrester <366@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1