Update patch set 1
Patch Set 1:

It's not a real icon or indicator so it has nothing to be consistent with, just the general appearance of the skin. But it could made the same as other icons if required.

A circled exclamation mark that interrupts text flow is unmissable no matter what colour, so it doesn't have to be the only orange thing on the page.

Patch-set: 1
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0