[BREAKING CHANGE] Use getDescription for title attribute and move to model
getDescription defines a tooltip-like description of a node, so
re-use this for the actual tooltip instead of duplicating logic.

Make getDescription an instance method of the model, rather than a
static method of view nodes/annotations, as we always need the model
attributes to generate it.

Breaking change: getDescription methods should be moved to the model
as instance methods.

Change-Id: If22a51970c35c4fade8f0b4246fa465fff526d00