[WIP] Fix pulling of annotations when typing at the beginning of the document
By allowing callers to disable part of
ve.dm.ElementLinearData.prototype.getRelativeOffset's initial direction
behaviour, which at least fixes the issue...
Specifically, this parameter controls whether or not getRelativeOffset
can force positive direction when the offset is 0 or less (even when we
specifically want negative direction), or force negative direction when the
offset is at/over the length of the data (even when we specifically want
positive direction - this doesn't change the bug, just makes the change more
logically consistent).
We may want to just get rid of the behaviour of forcing direction if the offset
is out of certain bounds completely, but that sounds like it might break other
things in obscure ways. See also I3090c258
The new parameter has to be sent through a few functions, and isn't documented
(yet... if we decide we actually want this change...)

Bug: 68597
Change-Id: I84870cc70a24c3639cee41769d2f826a26a19eba